Revert Fix for recent EMC screenname change
EMC has added a menu option to select what skin to use.
(The E2 Skin or the buildin EMC skin.)
The defealt option is set to the EMC-buildin skin, and therefor uses a new screenname.
Users are able to switch in the EMC settings, so there is no need for the earlier PR anymore.