Update sk.po
"InfoBar" and "InfoPanel" are two totally different views in Enigma2 ! The concept of "InfoPanel" is already being used in Enigma2 systems for another purpose ! It is therefore better if the original term is used - i.e. "InfoBar". Otherwise, the terms will be confusing. I returned this term to the original English "InfoBar", instead of the Slovak "InfoPanel".